Primera version de busqueda del objetivo
Posible falta de alguna condicion a la hora de busca o como saber que una solucion esta bien

+++ src/Nodo.java
@@ -3,11 +3,11 @@ import java.util.Comparator;
 public class Nodo{
 	private Nodo Padre; //Nodo padre
 	private Estado Est;	//Estado
-	private int Camino; //Camino recorrido
+	private double Camino; //Camino recorrido
 	private int P; //Profundidad a la que se encuentra el nodo
 	private int F;	//Valor en la frontera
 
-	public Nodo(Nodo Padre, Estado Est, int Camino, int P, int F) {
+	public Nodo(Nodo Padre, Estado Est, double Camino, int P, int F) {
 		this.Padre = Padre;
 		this.Est = Est;
 		this.Camino = Camino;
@@ -23,7 +23,7 @@ public class Nodo{
 		return Est;
 	}
 	
-	public int GetCamino() {
+	public double GetCamino() {
 		return Camino;
 	}
